Add workflow to leave PR comment
When a code snippet is updated often its line numbers also change. These line number references are used to render the snippet on ADP. It is easy to forget that step, and this Action will leave a reminder comment on the pull request to make that check prior to merging.